Fix for status bar changing to an html element.
The compose window status bar element was changed to an <html:div> in bug 1683865
(milestone 93). A quick search shows that other windows (esp. messenger.xhtml)
have not made the change yet.

Using just the class "statubar" to locate it should be sufficient and should work
in the cases where the statusbar is still an <hbox> as well as cases where it's
been changed.